update 5 files and delete 2 files
