+++ beakpeek/lib/Model/BirdInfo/bird_search_functions.dart
@@ -1,6 +1,6 @@
 import 'package:beakpeek/Model/BirdInfo/bird.dart';
 import 'package:beakpeek/Model/Globals/globals.dart';
-import 'package:beakpeek/Styles/global_styles.dart';
+import 'package:beakpeek/View/Home/Searching/filterable_searchbar_widgets.dart';
 import 'package:flutter/material.dart';
 
 final colorArray = [
@@ -20,41 +20,6 @@ int getColorForReportingRate(double reportingRate) {
   } else {
     return 3;
   }
-}
-
-Widget getData(Bird bird, Function goBird, BuildContext context) {
-  return ListTile(
-    title: Text(
-      bird.commonGroup.isNotEmpty
-          ? '${bird.commonGroup} ${bird.commonSpecies}'
-          : bird.commonSpecies,
-      style: GlobalStyles.filterTileHeading(context),
-    ),
-    subtitle: Text(
-      '${bird.genus} ${bird.species}',
-      style: GlobalStyles.filterTileSubHeading(context),
-    ),
-    trailing: Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        const SizedBox(width: 10),
-        FilledButton(
-          onPressed: () {
-            goBird(bird);
-          },
-          style: GlobalStyles.buttonFilterPrimaryFilled(context).copyWith(
-            elevation: WidgetStateProperty.all(2),
-          ),
-          child: Text('View Bird',
-              style: GlobalStyles.smallContentPrimary(context)
-                  .copyWith(color: Colors.white)),
-        ),
-      ],
-    ),
-    onTap: () {
-      goBird(bird);
-    },
-  );
 }
 
 bool isSeen(Bird bird) {
