+++ beakpeek/lib/View/Home/Searching/filterable_searchbar_widgets.dart
@@ -1,0 +1,38 @@
+import 'package:beakpeek/Model/BirdInfo/bird.dart';
+import 'package:beakpeek/Styles/global_styles.dart';
+import 'package:flutter/material.dart';
+
+Widget getData(Bird bird, Function goBird, BuildContext context) {
+  return ListTile(
+    title: Text(
+      bird.commonGroup.isNotEmpty
+          ? '${bird.commonGroup} ${bird.commonSpecies}'
+          : bird.commonSpecies,
+      style: GlobalStyles.filterTileHeading(context),
+    ),
+    subtitle: Text(
+      '${bird.genus} ${bird.species}',
+      style: GlobalStyles.filterTileSubHeading(context),
+    ),
+    trailing: Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        const SizedBox(width: 10),
+        FilledButton(
+          onPressed: () {
+            goBird(bird);
+          },
+          style: GlobalStyles.buttonFilterPrimaryFilled(context).copyWith(
+            elevation: WidgetStateProperty.all(2),
+          ),
+          child: Text('View Bird',
+              style: GlobalStyles.smallContentPrimary(context)
+                  .copyWith(color: Colors.white)),
+        ),
+      ],
+    ),
+    onTap: () {
+      goBird(bird);
+    },
+  );
+}
